# pims/pims/openapi/api-specification.yaml
openapi: 3.0.3
info:
  title: Cytomine Python Image Managament Server PIMS
  description: |
    Cytomine Python Image Managament Server (PIMS) HTTP API. While this API is intended to be internal, a lot of the
    following specification can be ported to the external Cytomine API.

    # Additional API formats
    In addition to OpenAPI types and formats, the following table contains some additional formats used in this
    specification.

    | Type | Format | Examples | Specification |
    |------|--------|----------|---------------|
    | `string` | `range` | **"1:3"**, **"2:"**, **":"**| Inspired from Python slicing |
    | `string` | `wkt` | **"POINT (10 20)"** | [Open Geospacial Consortium](https://www.ogc.org/standards/sfa)|
    | `string` | `color` | **"#FF00AA"**, **"rgb(100,0,0)"**, **"red"** | [CSS Color Module Level 3](https://drafts.csswg.org/css-color-3/#colorunits)|
    | `string` | `colormap-id` | **"jet"**, **"!jet"**||
    | `string` | `filter-id` | **"yen"** ||
  version: 0.1.0
#servers:
#  - url: 'https'
...
paths:
...
  /image/{filepath}/thumb:
    parameters:
      - $ref: '#/components/parameters/path-filepath'
      - $ref: '#/components/parameters/header-etag'
      - $ref: '#/components/parameters/header-x-image-size-safety'
    get:
      operationId: pims.api.thumb.show_thumb
      summary: Get a thumbnail
      description: |
        Get a 8-bit thumbnail optimized for visualisation, with given channels, focal planes and timepoints. If
        multiple channels are given (slice or selection), they are merged. If multiple focal planes or timepoints are
        given (slice or selection), a reduction function must be provided.

        **By default**, all image channels are used and when the image is multidimensional, the
        thumbnail is extracted from the median focal plane at first timepoint.
      tags:
        - Thumbnails
      parameters:
        - $ref: '#/components/parameters/query-channel-selection'
        - $ref: '#/components/parameters/query-z-slice-selection'
        - $ref: '#/components/parameters/query-timepoint-selection'
#        - $ref: '#/components/parameters/query-c-reduction'
#        - $ref: '#/components/parameters/query-z-reduction'
#        - $ref: '#/components/parameters/query-t-reduction'
        - $ref: '#/components/parameters/query-thumb-length'
        - $ref: '#/components/parameters/query-thumb-width'
        - $ref: '#/components/parameters/query-thumb-height'
        - $ref: '#/components/parameters/query-min-intensities-visualisation'
        - $ref: '#/components/parameters/query-max-intensities-visualisation'
#        - $ref: '#/components/parameters/query-colormaps'
        - $ref: '#/components/parameters/query-filters'
        - $ref: '#/components/parameters/query-gammas'
        - $ref: '#/components/parameters/query-log'
        - $ref: '#/components/parameters/query-thumb-use-precomputed'
      responses:
        200:
          $ref: '#/components/responses/thumbnail'
        304:
          $ref: '#/components/responses/not-modified'
        400:
          $ref: '#/components/responses/problem-bad-request'
        404:
          $ref: '#/components/responses/problem-not-found'
    post:
      operationId: pims.api.thumb.show_thumb_with_body
      summary: Get a thumbnail
      description: |
        **`GET with body` - when a GET with URL encoded query parameters is not possible due to URL size limits, a POST
        with body content must be used.**
        
        Get a 8-bit thumbnail optimized for visualisation, with given channels, focal planes and timepoints. If
        multiple channels are given (slice or selection), they are merged. If multiple focal planes or timepoints are
        given (slice or selection), a reduction function must be provided.

        **By default**, all image channels are used and when the image is multidimensional, the
        thumbnail is extracted from the median focal plane at first timepoint.
      tags:
        - Thumbnails
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - title: thumb-by-length
                  allOf:
                    - $ref: '#/components/schemas/target-length-object-required'
                    - $ref: '#/components/schemas/thumb-basis'
                - title: thumb-by-width
                  allOf:
                    - $ref: '#/components/schemas/target-width-object-required'
                    - $ref: '#/components/schemas/thumb-basis'
                - title: thumb-by-height
                  allOf:
                    - $ref: '#/components/schemas/target-height-object-required'
                    - $ref: '#/components/schemas/thumb-basis'
      responses:
        200:
          $ref: '#/components/responses/thumbnail'
        304:
          $ref: '#/components/responses/not-modified'
        400:
          $ref: '#/components/responses/problem-bad-request'
        404:
          $ref: '#/components/responses/problem-not-found'
  /image/{filepath}/resized:
    parameters:
      - $ref: '#/components/parameters/path-filepath'
      - $ref: '#/components/parameters/header-etag'
      - $ref: '#/components/parameters/header-x-image-size-safety'
    get:
      operationId: pims.api.resized.show_resized
      summary: Get a resized full image
      description: |
        Get a resized full image (thumbnail), with given channels, focal planes and timepoints. If
        multiple channels are given (slice or selection), they are merged. If multiple focal planes or timepoints are
        given (slice or selection), a reduction function must be provided.

        **By default**, all image channels are used and when the image is multidimensional, the
        thumbnail is extracted from the median focal plane at first timepoint.

        **While `/image/{filepath}/thumb` provides optimization for visualisation, this endpoint has a general purpose,
        such as computer vision, image processing or machine learning.**
      tags:
        - Resized
      parameters:
        - $ref: '#/components/parameters/query-channel-selection'
        - $ref: '#/components/parameters/query-z-slice-selection'
        - $ref: '#/components/parameters/query-timepoint-selection'
#        - $ref: '#/components/parameters/query-c-reduction'
#        - $ref: '#/components/parameters/query-z-reduction'
#        - $ref: '#/components/parameters/query-t-reduction'
        - $ref: '#/components/parameters/query-thumb-length'
        - $ref: '#/components/parameters/query-thumb-width'
        - $ref: '#/components/parameters/query-thumb-height'
        - $ref: '#/components/parameters/query-thumb-zoom'
        - $ref: '#/components/parameters/query-thumb-level'
        - $ref: '#/components/parameters/query-min-intensities'
        - $ref: '#/components/parameters/query-max-intensities'
#        - $ref: '#/components/parameters/query-colormaps'
        - $ref: '#/components/parameters/query-filters'
        - $ref: '#/components/parameters/query-gammas'
        - $ref: '#/components/parameters/query-log'
        - $ref: '#/components/parameters/query-bits'
        - $ref: '#/components/parameters/query-colorspace'
      responses:
        200:
          $ref: '#/components/responses/thumbnail'
        304:
          $ref: '#/components/responses/not-modified'
        400:
          $ref: '#/components/responses/problem-bad-request'
        404:
          $ref: '#/components/responses/problem-not-found'
        406:
          $ref: '#/components/responses/problem-not-acceptable'
    post:
      operationId: pims.api.resized.show_resized_with_body
      summary: Get a resized full image
      description: |
        **`GET with body` - when a GET with URL encoded query parameters is not possible due to URL size limits, a POST
        with body content must be used.**

        Get a resized full image (thumbnail), with given channels, focal planes and timepoints. If
        multiple channels are given (slice or selection), they are merged. If multiple focal planes or timepoints are
        given (slice or selection), a reduction function must be provided.

        **By default**, all image channels are used and when the image is multidimensional, the
        thumbnail is extracted from the median focal plane at first timepoint.

        **While `/image/{filepath}/thumb` provides optimization for visualisation, this endpoint has a general purpose,
        such as computer vision, image processing or machine learning.**
      tags:
        - Resized
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - title: resized-by-length
                  allOf:
                    - $ref: '#/components/schemas/target-length-object-required'
                    - $ref: '#/components/schemas/resized-basis'
                - title: resized-by-width
                  allOf:
                    - $ref: '#/components/schemas/target-width-object-required'
                    - $ref: '#/components/schemas/resized-basis'
                - title: resized-by-height
                  allOf:
                    - $ref: '#/components/schemas/target-height-object-required'
                    - $ref: '#/components/schemas/resized-basis'
                - title: resized-by-zoom
                  allOf:
                    - $ref: '#/components/schemas/target-zoom-object-required'
                    - $ref: '#/components/schemas/resized-basis'
                - title: resized-by-level
                  allOf:
                    - $ref: '#/components/schemas/target-level-object-required'
                    - $ref: '#/components/schemas/resized-basis'
      responses:
        200:
          $ref: '#/components/responses/thumbnail'
        304:
          $ref: '#/components/responses/not-modified'
        400:
          $ref: '#/components/responses/problem-bad-request'
        404:
          $ref: '#/components/responses/problem-not-found'
        406:
          $ref: '#/components/responses/problem-not-acceptable'
...
components:
  parameters:
    header-etag:
      name: ETag
      in: header
      description: |
        The RFC 7232 ETag header field in a response provides the entity-tag of
        a selected resource. The entity-tag is an opaque identifier for versions
        and representations of the same resource over time, regardless whether
        multiple versions are valid at the same time. An entity-tag consists of
        an opaque quoted string, possibly prefixed by a weakness indicator (see
        [RFC 7232 Section 2.3](https://tools.ietf.org/html/rfc7232#section-2.3)).
      required: false
      schema:
        type: string
    header-x-image-size-safety:
      name: X-Image-Size-Safety
      in: header
      description: |
        This header provides hints about the way the server has to deal with too large image responses.
        * `SAFE_REJECT` - Reject too large image response and throw a `400 Bad Request`.
        * `SAFE_RESIZE` - Resize the image response to an acceptable image size. Information about the resize are
        returned in `X-Image-Size-Limit` header.
        * `UNSAFE` - **At your own risk!** Try to fulfill the request but can cause unintended side effects
        (unreadable response, server slown down, server failure, ...). It should only be used in rare controlled
        situations.
      required: false
      schema:
        default: SAFE_RESIZE
        type: string
        enum:
          - SAFE_REJECT
          - SAFE_RESIZE
          - UNSAFE
...
    path-filepath:
      name: filepath
      in: path
      description: The file path, relative to server base path.
      required: true
      schema:
        type: string
        format: path
      example: '123/my-file.ext'
...
    query-channel-selection:
      name: channels
      in: query
      required: false
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/single-channel-index'
...
    query-z-slice-selection:
      name: z_slices
      in: query
      required: false
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/single-z-slice-index'
...
    query-timepoint-selection:
      name: timepoints
      in: query
      required: false
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/single-timepoint-index'
...
    query-c-reduction:
      name: c_reduction
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/channel-reduction'
    query-z-reduction:
      name: z_reduction
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/z-slice-reduction'
    query-t-reduction:
      name: t_reduction
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/timepoint-reduction'
    query-min-intensities-visualisation:
      name: min_intensities
      in: query
      required: false
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/min-intensity-selection-list-default-auto-image'
    query-max-intensities-visualisation:
      name: max_intensities
      in: query
      required: false
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/max-intensity-selection-list-default-auto-image'
    query-min-intensities:
      name: min_intensities
      in: query
      required: false
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/min-intensity-selection-list'
    query-max-intensities:
      name: max_intensities
      in: query
      required: false
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/max-intensity-selection-list'
    query-colormaps:
      name: colormaps
      in: query
      required: false
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/colormap-id-list'
    query-filters:
      name: filters
      in: query
      required: false
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/filter-id-list'
    query-gammas:
      name: gammas
      in: query
      required: false
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/gamma-list'
    query-log:
      name: log
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/log'
    query-bits:
      name: bits
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/bits'
    query-colorspace:
      name: colorspace
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/colorspace'
    query-thumb-length:
      name: length
      in: query
      description: |
        Length of the largest side of the thumbnail. The other dimension is adjusted to preserve the aspect ratio.

        **Ignored if other size-related parameter such as `width` or `height` is present.**
      required: false
      schema:
        $ref: '#/components/schemas/target-length-default-256'
    query-thumb-width:
      name: width
      in: query
      description: |
        Width of the thumbnail. Height is adjusted to preserve the aspect ratio.

        **Takes precedence over `length`.**
      required: false
      schema:
        $ref: '#/components/schemas/target-length'
    query-thumb-height:
      name: height
      in: query
      description: |
        Height of the thumbnail. Width is adjusted to preserve the aspect ratio.

        **Takes precedence over `width` and `length`.**
      required: false
      schema:
        $ref: '#/components/schemas/target-length'
    query-thumb-zoom:
      name: zoom
      in: query
      description: |
        The zoom level to consider as thumbnail.
        Zoom 0 has the worst resolution (smallest image, top of the image pyramid).
        Maximum admissible zoom level depends on image.

        **Takes precedence over `height`, `width` and `length`.**
      required: false
      schema:
        $ref: '#/components/schemas/zoom-or-level'
    query-thumb-level:
      name: level
      in: query
      description: |
        The tier level.
        Level 0 has the best resolution (largest image, basis of the image pyramid).
        Maximum admissible tier level depends on image.

        **Takes precedence over `zoom`, `height`, `width` and `length`.**
      required: false
      schema:
        $ref: '#/components/schemas/zoom-or-level'
    query-thumb-use-precomputed:
      name: use_precomputed
      in: query
      schema:
        $ref: '#/components/schemas/thumb-use-precomputed'
...
  schemas:
...
    single-channel-index:
      type: integer
      format: int32
      description: A single channel index. **By default**, all channels are considered.
      minimum: 0
...
    single-z-slice-index:
      type: integer
      format: int32
      description: A single focal plane index. **By default**, the median focal plane is considered.
      minimum: 0
...
    single-timepoint-index:
      type: integer
      format: int32
      description: A single timepoint index. **By default**, the first timepoint considered.
      minimum: 0
...
    channel-reduction:
      description: |
        Reduction function used to merge selected channels.
      type: string
      enum:
        - ADD
        - MIN
        - AVG
        - MAX
      default: ADD
...
    z-slice-reduction:
      description: |
        Reduction function used to combinate selected focal planes.

        **This parameter is required if `z_slices` is not a single focal plane index.**
      allOf:
        - $ref: '#/components/schemas/generic-reduction'
    timepoint-reduction:
      description: |
        Reduction function used to combinate selected timepoints.

        **This parameter is required if `timepoints` is not a single timepoint index.**
      allOf:
        - $ref: '#/components/schemas/generic-reduction'
...
    min-intensity-selection-list-default-auto-image:
      x-single-or-array: true
      x-default: AUTO_IMAGE
      allOf:
        - default: AUTO_IMAGE
        - $ref: '#/components/schemas/min-intensity-selection-list'
...
    max-intensity-selection-list-default-auto-image:
      x-single-or-array: true
      x-default: AUTO_IMAGE
      allOf:
        - default: AUTO_IMAGE
        - $ref: '#/components/schemas/max-intensity-selection-list'

    # Colormap
...
    colormap-id-list:
      x-single-or-array: true
      description: |
        A colormap is a function that maps the colors of the original image (source) to the colors of the response
        image. The usage of colormap produces a false-color representation of the original image and helps at
        understanding the image.

        Valid colormap names can be found with the endpoint `/colormaps`. The colormap can be reversed by prepending the
        colormap name with `!`.
      oneOf:
        - $ref: '#/components/schemas/colormap-id'
        - type: array
          description: |
            Expected format is an array of size:
            * 1 to apply the same colormap to all channels
            * equals to the number of channels used to render the response, in order to apply a colormap per channel.
          minItems: 1
          items:
            $ref: '#/components/schemas/colormap-id'
...
    filter-id-list:
#      x-single-or-array: true
      description: |
        An image filter is used to change the appareance of an image and helps at understanding the
        source image.

        Valid filter names can be found with the endpoint `/filters`.
      allOf:
        - $ref: '#/components/schemas/filter-id'
#        - type: array
#          description: |
#            Expected format is an array of size:
#            * 1 to apply the same filter to all channels
#            * equals to the number of channels used to render the response, in order to apply a filter per channel.
#          minItems: 1
#          items:
#            $ref: '#/components/schemas/filter-id'
...
    gamma-list:
      description: |
        Gamma performs a non-linear histogram adjustment. Pixel intensities in the original image are raised
        to the power of the gamma value.

        If `gamma < 1`, faint objects become more intense while bright objects do not.

        If `gamma > 1`, medium-intensity objects become fainter while bright objects do not.
      x-default: 1
#      x-single-or-array: true
      allOf:
        - allOf:
          - default: 1
          - $ref: '#/components/schemas/gamma'
#        - type: array
#          description: |
#            Expected format is an array of size:
#            * 1 to apply the gamma correction to all channels
#            * equals to the number of channels used to render the response in order to apply the gamma correction
#            per channel.
#          minItems: 1
#          items:
#            $ref: '#/components/schemas/gamma'
    log:
      type: boolean
      default: false
      description: |
        Apply a logarithmic scale on image data to ease observation of high dynamic range images such as 16-bit images.
...
    target-length:
      description: |
        Length of the largest side of the target. The other dimension is adjusted to preserve the aspect
        ratio.
      example: 256
      oneOf:
        - type: integer
          format: int64
          description: A length in pixels.
          minimum: 1
          example: 256
        - type: number
          format: double
          description: A length expressed in percentage relatively to the real image length.
          minimum: 0.0
          maximum: 1.0
          example: 0.5
    target-length-default-256:
      x-default: 256
      allOf:
        - default: 256
        - $ref: '#/components/schemas/target-length'
    target-length-object-required:
      type: object
      required:
        - length
      properties:
        length:
          $ref: '#/components/schemas/target-length'
...
    target-width-object-required:
      type: object
      required:
        - width
      properties:
        width:
          $ref: '#/components/schemas/target-width'
...
    target-height-object-required:
      type: object
      required:
        - height
      properties:
        height:
          $ref: '#/components/schemas/target-height'
...
    target-zoom-object-required:
      type: object
      required:
        - zoom
      properties:
        zoom:
          $ref: '#/components/schemas/target-zoom'
...
    target-level-object-required:
      type: object
      required:
        - level
      properties:
        level:
          $ref: '#/components/schemas/target-level'
...
    thumb-use-precomputed:
      description: Whether the precomputed thumbnail (associated file) has to be used if it is available, or not.
      type: boolean
      default: true
    thumb-basis:
      allOf:
        - $ref: '#/components/schemas/window-plane-selection'
        - $ref: '#/components/schemas/image-manipulation-visualisation'
        - type: object
          properties:
            use_precomputed:
              $ref: '#/components/schemas/thumb-use-precomputed'
    resized-basis:
      allOf:
        - $ref: '#/components/schemas/window-plane-selection'
        - $ref: '#/components/schemas/image-manipulation'
        - type: object
          properties:
            bits:
              $ref: '#/components/schemas/bits'
            colorspace:
              $ref: '#/components/schemas/colorspace'
...
    generic-image:
      type: string
      format: binary
      example: (binary)
...
  headers:
    If-None-Match:
      description: |
        The RFC7232 If-None-Match header field in a request requires the server
        to only operate on the resource if it does not match any of the provided
        entity-tags. If the provided entity-tag is `*`, it is required that the
        resource does not exist at all (see [RFC 7232 Section
        3.2](https://tools.ietf.org/html/rfc7232#section-3.2)).
      required: false
      schema:
        type: string
    Cache-Control:
      description: |
        The RFC 7234 Cache-Control header field is providing directives to
        control how proxies and clients are allowed to cache responses results
        for performance. Clients and proxies are free to not support caching of
        results, however if they do, they must obey all directives mentioned in
        [RFC-7234 Section 5.2.2](https://tools.ietf.org/html/rfc7234) to the
        word.

        In case of caching, the directive provides the scope of the cache
        entry, i.e. only for the original user (private) or shared between all
        users (public), the lifetime of the cache entry in seconds (max-age),
        and the strategy how to handle a stale cache entry (must-revalidate).
        Please note, that the lifetime and validation directives for shared
        caches are different (s-maxage, proxy-revalidate).
      required: false
      schema:
        type: string
        example: "private, must-revalidate, max-age=300"
    Vary:
      description: |
        The RFC 7231 Vary header field in a response defines which parts of
        a request message, aside the target URL and HTTP method, might have
        influenced the response. A client or proxy cache must respect this
        information, to ensure that it delivers the correct cache entry (see
        [RFC-7231 Section
        7.1.4](https://tools.ietf.org/html/rfc7231#section-7.1.4)).
      required: false
      schema:
        type: string
        example: "accept-encoding, accept-language"
...
    X-Image-Size-Limit:
      description: |
        This header indicates that the image representation has been resized due to reached image size limit. This
        limit can be set by the server or by the format itself.
      required: false
      style: simple
      explode: true
      schema:
        type: object
        properties:
          ratio:
            type: number
          request_width:
            type: integer
          request_height:
            type: integer
          safe_width:
            type: integer
          safe_height:
            type: integer
  responses:
    problem-bad-request:
      description: Bad request - Input payload fails business logic validation or generic error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/invalid-parameters-problem'
    problem-not-found:
      description: Not Found - The resource is not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/problem'
    problem-not-acceptable:
      description: |
        Not Acceptable - The resource can only generate content not acceptable according to the Accept
        headers sent in the request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/problem'
...
    not-modified:
      description: Not Modified
...
    thumbnail:
      description: OK - The thumbnail
      headers:
        If-None-Match:
          $ref: '#/components/headers/If-None-Match'
        Cache-Control:
          $ref: '#/components/headers/Cache-Control'
        Vary:
          $ref: '#/components/headers/Vary'
        X-Image-Size-Limit:
          $ref: '#/components/headers/X-Image-Size-Limit'
      content:
        image/jpeg:
          schema:
            $ref: '#/components/schemas/generic-image'
        image/png:
          schema:
            $ref: '#/components/schemas/generic-image'
        image/webp:
          schema:
            $ref: '#/components/schemas/generic-image'
...